gate linux-only imports
