+++ metricbeat/module/system/process/network/tracker_other.go
@@ -1,0 +1,36 @@
+// Licensed to Elasticsearch B.V. under one or more contributor
+// license agreements. See the NOTICE file distributed with
+// this work for additional information regarding copyright
+// ownership. Elasticsearch B.V. licenses this file to you under
+// the Apache License, Version 2.0 (the "License"); you may
+// not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+//go:build !linux
+
+package network
+
+import (
+	"context"
+	"errors"
+)
+
+// NewNetworkTracker returns an error on unsupported systems
+func NewNetworkTracker() (*Tracker, error) {
+	return nil, errors.New("network tracker is only available under Linux")
+}
+
+// Track is a bare tracker implementation
+func (track *Tracker) Track(ctx context.Context) error { return nil }
+
+// Stop is a bare stop implementation
+func (track *Tracker) Stop() {}
